pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "comattack"
version = "0.1.0"
description = "Black-box adversarial testing framework that exploits prompt compression as an attack surface."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Comattack Authors"},
]
keywords = ["adversarial", "prompt-compression", "llm", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.35",
    "numpy",
    "pyyaml",
    "tqdm",
]

[project.optional-dependencies]
llmlingua = ["llmlingua>=0.2"]
vllm = ["vllm>=0.4", "openai>=1.0"]
eval = ["datasets", "sentence-transformers", "nltk"]
all = ["comattack[llmlingua,vllm,eval]"]

[tool.setuptools.packages.find]
include = ["comattack*"]

[tool.setuptools.package-data]
"comattack.data.prompts" = ["*.txt"]
